+++ packages/react/src/experimental/OneDataCollection/__stories__/mockData.tsx
@@ -247,6 +247,7 @@ export const getMockVisualizations = (options?: {
     nestedRecords?: boolean
     applyLongText?: boolean
     longColumnLabels?: boolean
+    referenceRows?: boolean
   }
   cache?: MockDataCache<MockUser>
 }): Record<
@@ -269,6 +270,9 @@ export const getMockVisualizations = (options?: {
         allowColumnReordering: options?.table?.allowColumnReordering,
         frozenColumns:
           options?.table?.frozenColumns ?? options?.frozenColumns ?? 0,
+        referenceRowType: options?.table?.referenceRows
+          ? (item) => (item.status === "inactive" ? "striped" : "none")
+          : undefined,
         columns: [
           {
             label: "Name",